recommendations/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
blinker==1.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.0.0
pillow==10.4.0
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.1.0
uvicorn==0.32.1
Werkzeug==3.1.3
marshmallow
tmdbv3api==1.7.6
